Seller repository completed

+++ src/main/kotlin/com/example/models/Utils.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class SellerOpenStatus(
     val id: Int = 0,
-    val seller_id: Int = 0,
+    val seller_id: Long = 0,
     val is_open: Boolean = false
 )
